buffer can messages to avoid interrupting other code (#3)
* buffer can messages to avoid interrupting other code

Because CAN messages can arrive at any time, they might interrupt code in a critical section and cause reentrancy issues.

With this change, they will be handled by the main loop, with the interrupt handler only adding them to a buffer.

After discussion with Raffael Rott the code was moved to Can.cpp

* actually handle the can messages

* fix copying of bytes

* increase can buffer size

* add length method to ringbuf